A new menu system for views.
This is all the functionality needed for the page, app menus and browser system menus.

Review URL: http://codereview.chromium.org/119237

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@17895 0039d316-1c4b-4281-b951-d872f2087c98

+++ views/view.cc
@@ -943,7 +943,7 @@ void View::RemoveAccelerator(const Accelerator& accelerator) {
     return;
   }
 
-  int index = iter - accelerators_->begin();
+  size_t index = iter - accelerators_->begin();
   accelerators_->erase(iter);
   if (index >= registered_accelerator_count_) {
     // The accelerator is not registered to FocusManager.
